[#107] Store and use/set context for queued tasks
This patch makes the `enqueue` method store the context in which the
queue task was generated. This includes:

* interface
* userenv

This way the task queue will then use the newly introduced method
`execute_with_context` so the context is set accordingly for running the
task. It will be reset to its original state afterwards.

It is all covered by the added tests :-D

Signed-off-by: Tomás Cohen Arazi <tomascohen@theke.io>